+++ build.gradle
@@ -4,7 +4,7 @@ buildscript {
   ext {
     composeVersion = '1.0.0-alpha05'
     kotlinVersion = '1.4.10'
-    daggerVersion = '2.28.3'
+    daggerVersion = '2.29.1'
   }
 
   repositories {
