+++ lottie-compose/src/main/AndroidManifest.xml
@@ -1,5 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.airbnb.lottie.compose">
-
-</manifest>
+<manifest package="com.airbnb.lottie.compose" />
