+++ lottie/src/main/java/com/airbnb/lottie/LottieDrawable.java
@@ -13,7 +13,6 @@ import android.graphics.Typeface;
 import android.graphics.drawable.Animatable;
 import android.graphics.drawable.Drawable;
 import android.os.Build;
-import android.util.Log;
 import android.view.View;
 import android.widget.ImageView;
